Move transformers out of dimension package
