+++ server/storage/src/main/java/org/bithon/server/storage/datasource/dimension/LongDimensionSpec.java
@@ -18,7 +18,7 @@ package org.bithon.server.storage.datasource.dimension;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.bithon.server.storage.datasource.dimension.transformer.IDimensionTransformer;
+import org.bithon.server.storage.datasource.transformer.ITransformer;
 import org.bithon.server.storage.datasource.typing.IValueType;
 import org.bithon.server.storage.datasource.typing.LongValueType;
 
@@ -37,7 +37,7 @@ public class LongDimensionSpec extends AbstractDimensionSpec {
                              @JsonProperty("displayText") @NotNull String displayText,
                              @JsonProperty("required") @Nullable Boolean required,
                              @JsonProperty("visible") @Nullable Boolean visible,
-                             @JsonProperty("transformer") @Nullable IDimensionTransformer transformer) {
+                             @JsonProperty("transformer") @Nullable ITransformer transformer) {
         super(name, alias, displayText, required, visible, transformer);
     }
